requirements.txt
pytest
pytest_bdd
selenium==3.141
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest